Rearrange files
#1129
