+++ trview.lua.imgui/src/pch.h
@@ -10,4 +10,11 @@
 // add headers that you want to pre-compile here
 #include "framework.h"
 
+#include <external/lua/src/lua.h>
+#include <external/lua/src/lauxlib.h>
+
+#include <external/imgui/imgui.h>
+#include <external/imgui/imgui_internal.h>
+#include <external/imgui/misc/cpp/imgui_stdlib.h>
+
 #endif //PCH_H
