+++ trview.lua.imgui/src/tables.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <string>
+
+struct lua_State;
+
+namespace trview
+{
+    namespace lua
+    {
+        std::string get_string(lua_State* L, int index, const std::string& name);
+    }
+}
