+++ trview.lua.imgui/src/trview.lua.imgui.input.h
@@ -1,0 +1,11 @@
+#pragma once
+
+struct lua_State;
+
+namespace trview
+{
+    namespace lua
+    {
+        void register_input(lua_State* L);
+    }
+}
